Extract #underscore and #camelize into a dedicated mixin
